refactor: FSD-lite 폴더 구조로 리팩토링
