+++ frontend/src/01_shared/config/openLogAssets.ts
@@ -1,0 +1,14 @@
+export const openLogAssets = {
+  featuredCover:
+    "http://localhost:3845/assets/215944c0ae039b44468166b8e84bf99beae0f71d.png",
+  postCover:
+    "http://localhost:3845/assets/1d59d1970b8502f1a9dfc1f9aea722f8578e0946.png",
+  profileAvatar:
+    "http://localhost:3845/assets/1d59d1970b8502f1a9dfc1f9aea722f8578e0946.png",
+  knowledgeGraph:
+    "http://localhost:3845/assets/7cccabd62c9ceae10c4b94563846cc9452172b1c.png",
+  avatarA:
+    "http://localhost:3845/assets/8925e570ef6cac4accffa930abb18210e5de450e.png",
+  avatarB:
+    "http://localhost:3845/assets/97658fca33fb8976d18773f7d5ef527f8d3a91a5.png",
+} as const;
