+++ frontend/src/01_shared/config/openLogDemo.ts
@@ -1,0 +1,1 @@
+export const DEV_DEFAULT_IS_LOGGED_IN = false;
