+++ frontend/src/01_shared/lib/markdown/openLogMarkdownFormatting.ts
