+++ frontend/src/03_features/auth/ui/OpenLogGuestActions.tsx
