+++ frontend/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "baseUrl": ".",
     "target": "ES2017",
     "lib": ["dom", "dom.iterable", "esnext"],
     "allowJs": true,
@@ -19,7 +20,11 @@
       }
     ],
     "paths": {
-      "@/*": ["./*"]
+      "@/app/*": ["./app/*"],
+      "@/shared/*": ["./src/01_shared/*"],
+      "@/entities/*": ["./src/02_entities/*"],
+      "@/features/*": ["./src/03_features/*"],
+      "@/widgets/*": ["./src/04_widgets/*"]
     }
   },
   "include": [
